Update showcase.js
We have made a Companion app for our Blink smart watch in React Native. It will be launched in App Store soon. It is under review by apple team with our hardware.

If you want to know more about the company, You can checkout Mashable ( http://mashable.com/2016/09/30/blink-smartwatch-india/#aAdTEyTkfOqW ) , The Next Web ( http://thenextweb.com/gadgets/2016/09/30/the-blink-smartwatch-takes-the-fight-to-google-and-pebble-with-a-custom-os/ ) , Tech in Asia ( https://www.techinasia.com/witworks-blink-%3E%20smartwatch-apple-competitor ) etc., 

Thanks.,